refactor: Criação de funções service individualizadas.
