+++ app/Http/Controllers/AuthController.php
@@ -2,25 +2,29 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Services\Auth\AuthService;
+use App\Http\Requests\AuthRequest;
+use App\Http\Services\Auth\LoginService;
+use App\Http\Services\Auth\LogoutService;
 use Illuminate\Http\Request;
 
 class AuthController extends Controller
 {
-    private $authService;
+    private $loginService;
+    private $logoutService;
 
-    public function __construct(AuthService $authService)
+    public function __construct(LoginService $loginService, LogoutService $logoutService)
     {
-        $this->authService = $authService;
+        $this->loginService = $loginService;
+        $this->logoutService = $logoutService;
     }
 
     public function store(Request $request)
     {
-        return $this->authService->login($request);
+        return $this->loginService->login($request);
     }
 
     public function logout(Request $request)
     {
-        return $this->authService->logout($request);
+        return $this->logoutService->logout($request);
     }
 }
