+++ app/Http/Services/Auth/GetProfileService.php
@@ -1,0 +1,21 @@
+<?php
+
+namespace App\Http\Services\Auth;
+
+use App\Http\Repositories\AuthRepository;
+use Illuminate\Http\Request;
+
+class GetProfileService
+{
+    private $authRepository;
+    
+    public function __construct(AuthRepository $authRepository)
+    {
+        $this->authRepository = $authRepository;
+    }
+
+    public function getProfile(Request $request)
+    {
+        return $this->authRepository->findProfileById($request->user()->profile_id);
+    }
+}
