+++ app/Http/Services/Auth/TokenRevocationService.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Http\Services\Auth;
+
+use Illuminate\Http\Request;
+
+class TokenRevocationService
+{
+    public function revokeTokens(Request $request)
+    {
+        $request->user()->tokens()->delete();
+    }
+}
